New version of model. (and structure updates)
