+++ data/BP.csv
@@ -1,0 +1,61 @@
+Yhdiste,Nimi,Kiehumispiste_C°,Kiehumispiste_K,SMILES,Moolimassa
+1,Methane,-162.2,110.95,C,"16.043000
+"
+2,Ethane,-88.6,184.55,CC,"30.070000
+"
+3,propane,-42.2,230.95,CCC,"44.100000
+"
+4,butane,-0.1,273.05,CCCC,"58.120000
+"
+5,2-methylpropane,-11.2,261.95,CC(C)C,"58.120000
+"
+6,pentane,36.1,309.25,CCCCC,"72.150000
+"
+7,2-methylbutane,27,300.15,CC(C)CC,"72.150000
+"
+8,"2,2-dimethylpropane",9.5,282.65,CC(C)(C)C,"72.150000
+"
+9,hexane,68.8,341.95,CCCCCC,"86.180000
+"
+10,2-methylpentane,60.9,334.05,CC(C)CCC,"86.180000
+"
+11,3-methylpentane,63.3,336.45,CC(CC)CC,"86.180000
+"
+12,"2,2-dimethylbutane",49.8,322.95,CC(C)(CC)C,"86.180000
+"
+13,"2,3-dimethylbutane",58.1,331.25,CC(C)C(C)C,"86.180000
+"
+14,heptane,98.5,371.65,CCCCCCC,"100.200000
+"
+15,3-ethylpentane,93.5,366.65,C(C)C(CC)CC,"100.200000
+"
+16,"2,2-dimethylpentane",79.2,352.35,CC(C)(CCC)C,"100.200000
+"
+17,"2,3-dimethylpentane",89.8,362.95,CC(C)C(CC)C,"100.200000
+"
+18,"2,4-dimethylpentane",80.6,353.75,CC(C)CC(C)C,"100.200000
+"
+19,2-methylhexane,90.1,363.25,CC(C)CCCC,"100.205000
+"
+20,3-methylhexane,91.8,364.95,CC(CC)CCC,"100.200000
+"
+21,octane,125.6,398.75,CCCCCCCC,"114.230000
+"
+22,3-methylheptane,118.9,392.05,CC(CC)CCCC,"114.232000
+"
+23,"2,2,3,3-tetramethylbutane",106.5,379.65,CC(C)(C(C)(C)C)C,"114.230000
+"
+24,"2,3,3-trimethylpentane",114.7,387.85,CC(C)C(CC)(C)C,"114.230000
+"
+25,"2,3,4-trimethylpentane",113.7,386.85,CC(C)C(C(C)C)C,"114.230000
+"
+26,"2,2,4-trimethylpentane",99.3,372.45,CC(C)(CC(C)C)C,"114.230000
+"
+27,nonane,150.7,423.85,CCCCCCCCC,"128.250000
+"
+28,2-methyloctane,143,416.15,CC(C)CCCCCC,"128.259000
+"
+29,decane,174.2,447.35,CCCCCCCCCC,"142.280000
+"
+30,2-methylnonane,166.9,440.05,CC(C)CCCCCCC,"142.280000
+"
